ingestion/requirements.txt
beautifulsoup4
requests
deepl
python-dotenv
langchain-openai
mcp
schedule
